latest commit

+++ src/main/java/com/ascentdev/wims/entity/RackUtilEntity.java
@@ -58,8 +58,8 @@ public class RackUtilEntity {
   int modifiedById;
   
   @Column(name="txn_mawb_id")
-  long txnMawbId;
+  int txnMawbId;
   
   @Column(name="txn_hawb_id")
-  long txnHawbId;
+  int txnHawbId;
 }
